rename MySnackbarContentwrapper to Content and SnackbarContentWrapper to SnackbarContent
