+++ frontend/src/assest/styles/SnackBarContentStyle.js
